+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/user/UserStore.kt
@@ -1,0 +1,6 @@
+package com.sexysisters.tojserverv2.domain.user
+
+interface UserStore {
+    fun store(user: User): User
+    fun storeOAuthUser(user: User)
+}
