--- src/main/kotlin/com/sexysisters/tojserverv2/domain/user/design/UserStore.kt
@@ -1,8 +1,0 @@
-package com.sexysisters.tojserverv2.domain.user.design
-
-import com.sexysisters.tojserverv2.domain.user.User
-
-interface UserStore {
-    fun store(user: User): User
-    fun storeOAuthUser(user: User)
-}
