+++ src/main/kotlin/com/sexysisters/tojserverv2/infrastructure/user/UserReaderImpl.kt
@@ -2,7 +2,7 @@ package com.sexysisters.tojserverv2.infrastructure.user
 
 import com.sexysisters.tojserverv2.common.security.auth.AuthDetails
 import com.sexysisters.tojserverv2.domain.user.User
-import com.sexysisters.tojserverv2.domain.user.design.UserReader
+import com.sexysisters.tojserverv2.domain.user.UserReader
 import com.sexysisters.tojserverv2.domain.user.exception.UserException
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.security.core.context.SecurityContextHolder
