+++ src/test/kotlin/com/sexysisters/tojserverv2/domain/user/service/UserServiceTest.kt
@@ -4,8 +4,8 @@ import com.sexysisters.tojserverv2.config.properties.S3Properties
 import com.sexysisters.tojserverv2.domain.user.User
 import com.sexysisters.tojserverv2.domain.user.UserCommand
 import com.sexysisters.tojserverv2.domain.user.UserMapper
-import com.sexysisters.tojserverv2.domain.user.design.UserReader
-import com.sexysisters.tojserverv2.domain.user.design.UserStore
+import com.sexysisters.tojserverv2.domain.user.UserReader
+import com.sexysisters.tojserverv2.domain.user.UserStore
 import com.sexysisters.tojserverv2.domain.user.type.Authority
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
